requirements.txt
uuid == 1.30
cassandra-driver == 1.11.0
